backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]>=3.2.4,<4
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1,<5
python-multipart==0.0.17
email-validator==2.2.0
eval_type_backport>=0.2.0
